packages/python/algo_models/pyproject.toml
[build-system]
requires = ["maturin>=1,<2"]
build-backend = "maturin"

[project]
name = "algo_models"
version = "0.1.0"
requires-python = ">=3.7"
dependencies = ["pynacl>=1.5.0"]

[tool.maturin]
manifest-path = "../../../crates/algo_models_ffi/Cargo.toml"

[project.optional-dependencies]
dev = [
    'maturin>=1.8.2 ; platform_system != "Linux"',
    'maturin[patchelf]>=1.8.2 ; platform_system == "Linux"',
    "pynacl>=1.5.0",
    "pytest>=7.4.4",
]

[tool.pytest.ini_options]
markers = [
    "group_generic_transaction_tests",
    "group_transaction_tests"
]
